backend/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
python-multipart==0.0.6
opencv-python==4.7.0.72
pillow==9.5.0
torch==2.0.0
torchvision==0.15.1
scipy==1.10.1
numpy==1.24.3
matplotlib==3.7.1
pydantic==1.10.7
python-jose==3.3.0
